add lib/processing/context.rb
