+++ lib/processing.rb
@@ -17,4 +17,5 @@ require 'processing/shader'
 require 'processing/capture'
 require 'processing/graphics_context'
 require 'processing/graphics'
+require 'processing/context.rb'
 require 'processing/processing'
